# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0", 
      "title": "My Lab3 KPP Menu Project",
      "description": "API S",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:4000",
    "basePath": "/",
    "tags": [
      {
        "name": "Dishesh",
        "description": "API for dieshesh in the menu."
      }
    ],
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
        "/dishesh": {
          "post": {
            "operationId": "POST_dishesh",
            "produces": [
              "application/json"
            ],
            "responses": {
              "200": {
                "description": "",
                "schema": {
                  "example": {
                    "name": "soup",
                    "description": "hello",
                    "price": "10",
                    "imageURL": "https://sfllgffl",
                    "_id": "60d06c0694fe6c1c886964b7"
                  }
                }
              }
            },
            "consumes": [
              "application/json"
            ],
            "parameters": [
              {
                "schema": {
                  "example": {
                    "name": "Dish Name",
                    "description": "Dish Description",
                    "price": 10.4,
                    "imageURL": "https://image.jpg"
                  },
                  "type": "string"
                },
                "in": "body",
                "name": "body",
                "required": true
              }
            ]
          },
          "get": {
            "operationId": "GET_dishesh",
            "produces": [
              "application/json"
            ],
            "responses": {
              "200": {
                "description": "",
                "schema": {
                  "example": [
                    {
                      "_id": "60cdfcdd0535d50dc41d29f2",
                      "name": "Wine",
                      "description": "Spain",
                      "price": "200",
                      "imageURL": "https://static01.nyt.com/images/2021/04/01/dining/01next1/merlin_185496183_777290b3-7f9e-42b9-acee-0b1c3a1b3d4d-superJumbo.jpg"
                    },
                    {
                      "_id": "60ce35f509281624083bee72",
                      "name": "soup",
                      "description": "test",
                      "price": "11",
                      "imageURL": "https://imagesvc.meredithcorp.io/v3/mm/image?url=https%3A%2F%2Fstatic.onecms.io%2Fwp-content%2Fuploads%2Fsites%2F44%2F2020%2F01%2F02%2F7551492.jpg"
                    }
                  ]
                }
              }
            }
          }
        },
        "/dishesh/{dishId}": {
          "delete": {
            "operationId": "DELETE_dishesh-dishId",
            "produces": [
              "application/json"
            ],
            "responses": {
              "200": {
                "description": "",
                "schema": {
                  "example": [
                    {
                      "_id": "60cdfcdd0535d50dc41d29f2",
                      "name": "Wine",
                      "description": "Spain",
                      "price": "200",
                      "imageURL": "https://static01.nyt.com/images/2021/04/01/dining/01next1/merlin_185496183_777290b3-7f9e-42b9-acee-0b1c3a1b3d4d-superJumbo.jpg"
                    },
                    {
                      "_id": "60ce35f509281624083bee72",
                      "name": "soup",
                      "description": "test",
                      "price": "11",
                      "imageURL": "https://imagesvc.meredithcorp.io/v3/mm/image?url=https%3A%2F%2Fstatic.onecms.io%2Fwp-content%2Fuploads%2Fsites%2F44%2F2020%2F01%2F02%2F7551492.jpg"
                    }
                  ]
                }
              }
            }
          },
          "put": {
            "operationId": "PUT_dishesh-dishId",
            "produces": [
              "application/json"
            ],
            "responses": {
              "200": {
                "description": ""
              }
            }
          },
          "parameters": [
            {
              "type": "string",
              "in": "path",
              "name": "dishId",
              "required": true
            }
          ]
        }
      }
  }